working with game API
