+++ server.js
@@ -3,6 +3,10 @@ const session = require('express-session');
 const routes = require('./controllers');
 const path = require('path')
 const exphbs = require('express-handlebars')
+// const axios = require('axios');
+
+
+
 
 const helpers = require('./utils/helpers');
 const sequelize = require('./config/connection');
